fix: refactor network card
